Code to implement Dry Run functionality.

+++ loader/readers/base_reader.py
@@ -16,7 +16,7 @@ class BaseReader:
     this base class and implement the read_header() and parse_line() methods.
     """
     
-    def __init__(self, pattern, posters, id_to_bmon={}, default_bmon=None, 
+    def __init__(self, pattern, posters, dry_run=False, id_to_bmon={}, default_bmon=None, 
                 chunk_size=300, time_zone='US/Alaska', file_retention=3, **kw):
         """Constructor for BaseReader.
 
@@ -25,6 +25,11 @@ class BaseReader:
           posters: a dictionary of HttpPoster objects that are used to post to BMON
               servers.  The dictionary is keyed on BMON id values that appear in the
               bmon_servers section of the config file.
+          dry_run: If True, sensor readings will not be posted to the BMON site; instead
+              they will be appended to a file placed in a 'debug' subdirectory of the 
+              directory containing the files being processed.  The name of the file will
+              be '{BMON id}.txt'. Also, when 'dry_run' is True, no files will be deleted,
+              including the processed files and old files in the completed directory.
           id_to_bmon: a dictionary that maps sensors ID values to the destination BMON
               system, which is identified with a BMON id.
           default_bmon: the ID of the BMON system where a sensor reading should go if
@@ -48,6 +53,7 @@ class BaseReader:
         # Store most of the parameters as object attributes
         self.pattern = pattern
         self.posters = posters
+        self.dry_run = dry_run
         self.id_to_bmon = id_to_bmon
         self.default_bmon = default_bmon
         self.chunk_size = chunk_size
@@ -65,13 +71,19 @@ class BaseReader:
         # and error readings.
         (self.file_dir / 'completed').mkdir(exist_ok=True)
         (self.file_dir / 'errors').mkdir(exist_ok=True)
-        (self.file_dir / 'debug').mkdir(exist_ok=True)
+        (self.file_dir / 'debug').mkdir(exist_ok=True)  # for dry run files
 
         # Clean out old files in completed directory
-        max_age = file_retention * 24. * 3600.    # seconds
-        for f_path in (self.file_dir / 'completed').glob('*'):
-            if time.time() - f_path.stat().st_mtime > max_age:
-                f_path.unlink()
+        if not dry_run:
+            max_age = file_retention * 24. * 3600.    # seconds
+            for f_path in (self.file_dir / 'completed').glob('*'):
+                if time.time() - f_path.stat().st_mtime > max_age:
+                    f_path.unlink()
+
+        # If a Dry Run, clean out the files in the Debug directory
+        if dry_run:
+            for p in (self.file_dir / 'debug').glob('*'):
+                p.unlink()
 
     def load(self):
         """Called to start the processing of files.
@@ -91,12 +103,12 @@ class BaseReader:
 
                 if len(buf) >= min_post_size:
 
-                    self.posters[bmon_id].add_readings(buf)
-
-                    if logging.root.level == logging.DEBUG:
+                    if self.dry_run:
                         with open(self.file_dir / 'debug' / f'{bmon_id}.txt', 'a') as fout:
                             for rd in buf:
                                 print(rd, file=fout)
+                    else:
+                        self.posters[bmon_id].add_readings(buf)
                     
                     rd_buffer[bmon_id] = []  # reset buffer
 
@@ -181,8 +193,9 @@ class BaseReader:
 
             logging.info(f'Processed {fn}: {success_ct} successful lines, {error_ct} error lines.')
             
-            # delete the processed file
-            Path(fn).unlink()
+            # delete the processed file, unless it is a dry run
+            if not self.dry_run:
+                Path(fn).unlink()
 
         # Send remaining readings to BMON poster
         post_buffer()
